requirements.txt
kafka-python==2.0.2
PyYAML==6.0
python-dateutil==2.8.2
faker==18.4.0
pytest==7.3.1
python-dotenv==1.0.0
confluent-kafka==2.3.0 